requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python
numpy
pandas
Pillow
sounddevice
scikit-learn
tqdm
